@@ -1,3 +1,3 @@
 ﻿@typeparam TItem
 
-<div id="@Id" contenteditable="true" class=@ActiveClass style="@ComputedStyle" @onblur="HasLostFocus" @onfocus="HasGotFocus" @attributes="AdditionalAttributes" />
+<div id="@Id" contenteditable="true" class=@ActiveClass style="@ComputedStyle" @onblur="HasLostFocus" @onfocus="HasGotFocus" @attributes="AdditionalAttributes" inputmode="numeric" />
